Renamed all packages and maven from (de|com).zeiss to at.bestsolution
