+++ maven-osgi-targetplatform-extension/pom.xml
@@ -2,7 +2,7 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>com.zeiss.maven</groupId>
+	<groupId>at.bestsolution</groupId>
 	<artifactId>maven-osgi-targetplatform-extension</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
@@ -22,7 +22,7 @@
 	<dependencies>
 
 		<dependency>
-			<groupId>com.zeiss.maven</groupId>
+			<groupId>at.bestsolution</groupId>
 			<artifactId>maven-osgi-targetplatform-lib</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
