+++ maven-osgi-targetplatform-lib/src/main/java/at/bestsolution/maven/osgi/targetplatform/lib/ParameterProvider.java
@@ -1,0 +1,23 @@
+package at.bestsolution.maven.osgi.targetplatform.lib;
+
+
+/**
+ * Provides all necessary parameters.
+ * 
+ *
+ */
+public interface ParameterProvider {
+
+    String getAdditionalDependenciesFile();
+
+    String getWhitelistFile();
+
+    String getFeatureFile();
+
+    String getTargetFeatureJarPrefix();
+
+    String getEfxclipseSite();
+
+    String getEfxclipseUpdateSite();
+
+}
