--- maven-osgi-targetplatform-lib/src/main/java/de/zeiss/maven/osgi/targetplatform/lib/ParameterProvider.java
@@ -1,23 +1,0 @@
-package de.zeiss.maven.osgi.targetplatform.lib;
-
-
-/**
- * Provides all necessary parameters.
- * 
- *
- */
-public interface ParameterProvider {
-
-    String getAdditionalDependenciesFile();
-
-    String getWhitelistFile();
-
-    String getFeatureFile();
-
-    String getTargetFeatureJarPrefix();
-
-    String getEfxclipseSite();
-
-    String getEfxclipseUpdateSite();
-
-}
