+++ maven-osgi-targetplatform-plugin/pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.zeiss.maven</groupId>
+	<groupId>at.bestsolution</groupId>
 	<artifactId>maven-osgi-targetplatform-plugin</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>maven-plugin</packaging>
@@ -24,7 +24,7 @@
 	<dependencies>
 
 		<dependency>
-			<groupId>sample.mvn</groupId>
+			<groupId>at.bestsolution</groupId>
 			<artifactId>maven-osgi-targetplatform-lib</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
