--- maven-osgi-targetplatform-plugin/src/main/java/de/zeiss/maven/osgi/targetplatform/plugin/internal/ExtendedParameterProvider.java
@@ -1,19 +1,0 @@
-package de.zeiss.maven.osgi.targetplatform.plugin.internal;
-
-import de.zeiss.maven.osgi.targetplatform.lib.ParameterProvider;
-
-/**
- * Extends the parameter provider with parameters only needed by the plugin.
- *
- */
-public interface ExtendedParameterProvider extends ParameterProvider {
-
-    String getVersion();
-
-    String getArtifactId();
-
-    String getGroupId();
-
-    String getOutputFile();
-
-}
